CUBBONUS

+++ Srcs/Display/Dis_walls/dis_walls_utils.c
@@ -6,7 +6,7 @@
 /*   By: abrun <abrun@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/25 16:46:48 by abrun             #+#    #+#             */
-/*   Updated: 2022/01/14 17:06:39 by abrun            ###   ########.fr       */
+/*   Updated: 2022/01/31 12:34:28 by abrun            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ double	get_angle_right(int angle_h, double counter)
 	double	angle;
 
 	angle = angle_h + counter;
-	if (angle >= 360)
+	if ((int)angle >= 360)
 		angle -= 360;
 	if ((int)angle < 0)
 		angle = 360 + angle;
